# docs/architecture/api/openapi.yaml
openapi: 3.1.0
info:
  title: CWMS Construction Work Management System API
  version: 1.0.0
  summary: Implementation-agnostic HTTP API for CWMS v1.0 web application
  description: |
    Architecture-level OpenAPI 3.1 specification for CWMS Version 1.0.
    Not tied to any framework. Traceability to Functional Specification is via
    `x-fs-refs` on operations and the companion API-CATALOGUE.md.
    Money amounts are decimal strings. Dates are ISO-8601.
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Version 1 API base path
...
security:
  - sessionAuth: []
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: authLogin
      x-fs-refs: [F-AUTH-01, F-AUTH-01A, F-AUTH-01B, PRD-AUTH]
      x-prd-refs: [PRD-AUTH]
      x-br-refs: [BR-SEC-01, BR-SEC-03]
      security: []
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Inactive account
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: authLogout
      x-fs-refs: [F-AUTH-03]
      summary: Sign out
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      operationId: authMe
      x-fs-refs: [G-01]
      summary: Current user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/change-password:
    post:
      tags: [Auth]
      operationId: authChangePassword
      x-fs-refs: [F-AUTH-02]
      x-br-refs: [BR-SEC-02]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: authRefresh
      x-fs-refs: [F-AUTH-01A]
      summary: Refresh session / Remember Me
      responses:
        '200':
          description: Session extended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /bills:
    get:
      tags: [Billing]
      operationId: listBills
      x-fs-refs: [F-BILL-01]
      x-prd-refs: [PRD-BILL]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - name: q
          in: query
          schema: { type: string }
        - name: workId
          in: query
          schema: { type: string }
        - name: paymentStatus
          in: query
          schema: { $ref: '#/components/schemas/PaymentStatus' }
        - name: billType
          in: query
          schema: { $ref: '#/components/schemas/BillType' }
        - name: billDateFrom
          in: query
          schema: { type: string, format: date }
        - name: billDateTo
          in: query
          schema: { type: string, format: date }
        - name: financialYear
          in: query
          schema: { type: string }
        - name: client
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Bill page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillPage'
    post:
      tags: [Billing]
      operationId: createBill
      x-fs-refs: [F-BILL-02, F-FIN-03]
      x-br-refs: [BR-BILL-01, BR-BILL-02, BR-BILL-04, BR-BILL-07, BR-BILL-08, BR-ID-02, BR-FIN-08, BR-FIN-09, BR-FIN-15]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillCreateRequest'
      responses:
        '201':
          description: Created (work financials recalculated in same transaction)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /bills/{billId}:
    parameters:
      - $ref: '#/components/parameters/BillId'
    get:
      tags: [Billing]
      operationId: getBill
      x-fs-refs: [F-BILL-06]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Billing]
      operationId: updateBill
      x-fs-refs: [F-BILL-03, F-BILL-05]
      x-br-refs: [BR-BILL-10, BR-FIN-15]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Billing]
      operationId: deleteBill
      x-fs-refs: [F-BILL-04]
      x-br-refs: [BR-BILL-16, BR-FIN-15]
      responses:
        '204':
          description: Deleted and work rollups recalculated
        '403':
          $ref: '#/components/responses/Forbidden'
  /works/{workId}/bills:
    get:
      tags: [Billing]
      operationId: listWorkBills
      x-fs-refs: [F-BILL-06]
      parameters:
        - $ref: '#/components/parameters/WorkId'
      responses:
        '200':
          description: Bill history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillList'
...
  /reports/{reportType}/run:
    post:
      tags: [Reports]
      operationId: runReport
      x-fs-refs: [F-RPT-01]
      x-br-refs: [BR-RPT-01, BR-RPT-03, BR-RPT-08]
      parameters:
        - $ref: '#/components/parameters/ReportType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRunRequest'
      responses:
        '200':
          description: Tabular/sectioned JSON result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /reports/{reportType}/export:
    post:
      tags: [Reports]
      operationId: exportReport
      x-fs-refs: [F-RPT-03, F-RPT-04]
      x-br-refs: [BR-RPT-02, BR-RPT-04, BR-RPT-05]
      parameters:
        - $ref: '#/components/parameters/ReportType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportExportRequest'
      responses:
        '200':
          description: File download
          content:
            application/pdf:
              schema:
                type: string
                contentMediaType: application/pdf
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                contentMediaType: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
        '400':
          $ref: '#/components/responses/BadRequest'
  /reports/{reportType}/saved-filters:
    parameters:
      - $ref: '#/components/parameters/ReportType'
    get:
      tags: [Reports]
      operationId: listSavedFilters
      x-fs-refs: [F-RPT-05]
      x-br-refs: [BR-RPT-06]
      responses:
        '200':
          description: Saved filters for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedReportFilterList'
    post:
      tags: [Reports]
      operationId: createSavedFilter
      x-fs-refs: [F-RPT-05]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedReportFilterCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedReportFilter'
  /reports/{reportType}/saved-filters/{filterId}:
    parameters:
      - $ref: '#/components/parameters/ReportType'
      - $ref: '#/components/parameters/FilterId'
    patch:
      tags: [Reports]
      operationId: updateSavedFilter
      x-fs-refs: [F-RPT-05]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedReportFilterUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedReportFilter'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Reports]
      operationId: deleteSavedFilter
      x-fs-refs: [F-RPT-05]
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /backups:
    get:
      tags: [Backups]
      operationId: listBackups
      x-fs-refs: [F-BAK-02]
      x-br-refs: [BR-BAK-04]
      x-prd-refs: [PRD-BAK]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/BackupStatus' }
      responses:
        '200':
          description: Backup history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupPage'
        '403':
          $ref: '#/components/responses/Forbidden'
  /backups/{backupId}:
    get:
      tags: [Backups]
      operationId: getBackup
      x-fs-refs: [F-BAK-02]
      parameters:
        - $ref: '#/components/parameters/BackupId'
      responses:
        '200':
          description: Backup detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupRecord'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /backups/{backupId}/restore:
    post:
      tags: [Backups]
      operationId: restoreBackup
      x-fs-refs: [F-BAK-03]
      x-br-refs: [BR-BAK-05, BR-BAK-06, BR-BAK-07, BR-BAK-09]
      parameters:
        - $ref: '#/components/parameters/BackupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRestoreRequest'
      responses:
        '202':
          description: Restore accepted / completed (implementation may be sync or async)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupRestoreResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          description: Maintenance / restore conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: CWMSSESSION
      description: Server session cookie (implementation-defined name)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer/session token alternative
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      name: sort
      in: query
      description: field:asc|desc[,field:asc|desc]
      schema: { type: string }
...
    WorkId:
      name: workId
      in: path
      required: true
      schema: { type: string }
...
    BillId:
      name: billId
      in: path
      required: true
      schema: { type: string }
...
    ReportType:
      name: reportType
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ReportType'
    FilterId:
      name: filterId
      in: path
      required: true
      schema: { type: string }
    BackupId:
      name: backupId
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Not authenticated
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Not authorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    Money:
      type: string
      pattern: '^-?\d+(\.\d{1,2})?$'
      description: Decimal money as string
      example: "1000000.00"
...
    Role:
      type: string
      enum: [Administrator, DataEntryOperator, Engineer, Accounts, Viewer]
...
    BillType:
      type: string
      enum: [RaBill, FinalBill]
    PaymentStatus:
      type: string
      enum: [Pending, PartiallyReceived, FullyReceived]
...
    ReportType:
      type: string
      enum:
        - work-register
        - billing
        - expenditure
        - financial-summary
        - work-wise-summary
        - pending-payment
        - document-register
        - general-expense
        - dashboard-summary
    BackupStatus:
      type: string
      enum: [Success, Failed, Running]
    ExportFormat:
      type: string
      enum: [pdf, excel]
    ProblemDetails:
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Machine-stable error code
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              code: { type: string }
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
        rememberMe: { type: boolean, default: false }
    LoginResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/User' }
        token:
          type: string
          description: Optional bearer token if not using cookie-only sessions
        expiresAt: { type: string, format: date-time }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword, confirmNewPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 8 }
        confirmNewPassword: { type: string, minLength: 8 }
    User:
      type: object
      required: [id, name, loginId, role, active]
      properties:
        id: { type: string }
        name: { type: string }
        loginId: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        mobile: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    BillDeductionLine:
      type: object
      required: [name, amount]
      properties:
        id: { type: string }
        name: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        kind:
          type: string
          enum: [Standard, Other]
    Bill:
      type: object
      required: [id, workId, systemBillNumber, billType, billDate, grossBillAmount, totalDeductions, netBillAmount, paymentStatus]
      properties:
        id: { type: string }
        workId: { type: string }
        workCode: { type: string }
        workName: { type: string }
        systemBillNumber: { type: string, example: BILL-2026-0001 }
        billType: { $ref: '#/components/schemas/BillType' }
        raBillNo: { type: string, nullable: true }
        billDate: { type: string, format: date }
        periodFrom: { type: string, format: date, nullable: true }
        periodTo: { type: string, format: date, nullable: true }
        previousBillAmount: { $ref: '#/components/schemas/Money' }
        currentWorkPortionAmount: { $ref: '#/components/schemas/Money' }
        gstAmount: { $ref: '#/components/schemas/Money' }
        grossBillAmount: { $ref: '#/components/schemas/Money' }
        deductions:
          type: array
          items: { $ref: '#/components/schemas/BillDeductionLine' }
        totalDeductions: { $ref: '#/components/schemas/Money' }
        netBillAmount: { $ref: '#/components/schemas/Money' }
        paymentStatus: { $ref: '#/components/schemas/PaymentStatus' }
        paymentDate: { type: string, format: date, nullable: true }
        amountReceived: { $ref: '#/components/schemas/Money' }
        outstandingAmount: { $ref: '#/components/schemas/Money' }
        utrChequeNo: { type: string, nullable: true }
        bankName: { type: string, nullable: true }
        remarks: { type: string, nullable: true }
    BillCreateRequest:
      type: object
      required: [workId, billType, billDate, currentWorkPortionAmount, gstAmount, paymentStatus]
      properties:
        workId: { type: string }
        billType: { $ref: '#/components/schemas/BillType' }
        raBillNo: { type: string }
        billDate: { type: string, format: date }
        periodFrom: { type: string, format: date }
        periodTo: { type: string, format: date }
        previousBillAmount: { $ref: '#/components/schemas/Money' }
        currentWorkPortionAmount: { $ref: '#/components/schemas/Money' }
        gstAmount: { $ref: '#/components/schemas/Money' }
        standardDeductions:
          type: object
          additionalProperties: { $ref: '#/components/schemas/Money' }
          description: Map of standard deduction head name to amount
        otherDeductions:
          type: array
          items: { $ref: '#/components/schemas/BillDeductionLine' }
        paymentStatus: { $ref: '#/components/schemas/PaymentStatus' }
        paymentDate: { type: string, format: date }
        amountReceived: { $ref: '#/components/schemas/Money' }
        utrChequeNo: { type: string }
        bankName: { type: string }
        remarks: { type: string }
    BillUpdateRequest:
      $ref: '#/components/schemas/BillCreateRequest'
    BillList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Bill' }
    BillPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Bill' }
        page: { $ref: '#/components/schemas/PageMeta' }
...
    ReportRunRequest:
      type: object
      required: [filters]
      properties:
        filters:
          type: object
          additionalProperties: true
          description: Report-specific filter map (see Reporting Catalogue)
        sort: { type: string }
    ReportRunResponse:
      type: object
      required: [reportType, columns, rows, generatedAt]
      properties:
        reportType: { $ref: '#/components/schemas/ReportType' }
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        totals:
          type: object
          additionalProperties: true
        sections:
          type: array
          description: For sectioned reports e.g. financial-summary / work-wise-summary
          items:
            type: object
            additionalProperties: true
        generatedAt: { type: string, format: date-time }
        filtersApplied:
          type: object
          additionalProperties: true
    ReportExportRequest:
      allOf:
        - $ref: '#/components/schemas/ReportRunRequest'
        - type: object
          required: [format]
          properties:
            format: { $ref: '#/components/schemas/ExportFormat' }
    SavedReportFilter:
      type: object
      required: [id, reportType, name, filters, isDefault]
      properties:
        id: { type: string }
        reportType: { $ref: '#/components/schemas/ReportType' }
        name: { type: string }
        filters:
          type: object
          additionalProperties: true
        isDefault: { type: boolean }
        updatedAt: { type: string, format: date-time }
    SavedReportFilterCreateRequest:
      type: object
      required: [name, filters]
      properties:
        name: { type: string, minLength: 1 }
        filters:
          type: object
          additionalProperties: true
        isDefault: { type: boolean, default: false }
    SavedReportFilterUpdateRequest:
      type: object
      properties:
        name: { type: string }
        filters:
          type: object
          additionalProperties: true
        isDefault: { type: boolean }
    SavedReportFilterList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SavedReportFilter' }
...
    BackupRecord:
      type: object
      required: [id, startedAt, type, status]
      properties:
        id: { type: string }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time, nullable: true }
        type:
          type: string
          enum: [Automatic, Initial]
        status: { $ref: '#/components/schemas/BackupStatus' }
        identifier: { type: string }
        retainUntil: { type: string, format: date-time, nullable: true }
        message: { type: string, nullable: true }
    BackupPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BackupRecord' }
        page: { $ref: '#/components/schemas/PageMeta' }
    BackupRestoreRequest:
      type: object
      required: [confirmPhrase, acknowledgedDestructive]
      properties:
        confirmPhrase:
          type: string
          description: Must equal RESTORE
          enum: [RESTORE]
        acknowledgedDestructive: { type: boolean }
    BackupRestoreResponse:
      type: object
      required: [backupId, status]
      properties:
        backupId: { type: string }
        status:
          type: string
          enum: [Completed, Accepted, Failed]
        message: { type: string }
...